feat: Tests for i64x2 arithmetic operations...
...  on major boundary values and all special values.

Added 3 new instructions: i64x2.neg, i64x2.sub and i64x2.mul.

+++ WebAssembly/Instructions/Int64X2Add.cs
@@ -7,7 +7,7 @@ namespace WebAssembly.Instructions;
 /// </summary>
 public class Int64X2Add : SimdValueTwoToOneInstruction
 {
-    private static readonly MethodInfo addMethod = FindVector128Method("Add", typeof(uint));
+    private static readonly MethodInfo addMethod = FindVector128Method("Add", typeof(ulong));
 
     /// <summary>
     /// Always <see cref="SimdOpCode.Int64X2Add"/>.
